feat(twitter.function): リファクタ、ランダム処理を修正
